refactor(research): extract OCR asset quality tests
